Add authorization to expenses routes

+++ backend/src/controllers/expenseController.ts
@@ -1,24 +1,51 @@
+import { AuthRequest } from "./../middleware/authMiddleware";
 import { NextFunction, Request, Response } from "express";
 import { Expense } from "../models/expenseModel";
 
 const ExpenseController = () => {
-	const index = async (req: Request, res: Response, next: NextFunction) => {
-		const results = await Expense.find({});
-		res.status(200).send(results);
+	const index = async (
+		req: AuthRequest,
+		res: Response,
+		next: NextFunction
+	) => {
+		try {
+			if (req.user) {
+				const results = await Expense.find({ user: req.user.id });
+				res.status(200).send(results);
+			} else {
+				res.status(400);
+				throw new Error("User invalid");
+			}
+		} catch (error) {
+			next(error);
+		}
 	};
 
-	const create = async (req: Request, res: Response, next: NextFunction) => {
-		console.log(req.body);
-		if (!req.body.name) {
-			res.status(400);
-			throw new Error("Please add the name of the expense");
+	const create = async (
+		req: AuthRequest,
+		res: Response,
+		next: NextFunction
+	) => {
+		try {
+			if (req.user) {
+				if (!req.body.name) {
+					res.status(400);
+					throw new Error("Please add the name of the expense");
+				}
+				const result = await Expense.create({
+					name: req.body.name,
+					quantity: req.body.quantity,
+					description: req.body.description,
+					user: req.user.id,
+				});
+				res.status(200).send(result);
+			} else {
+				res.status(400);
+				throw new Error("User invalid");
+			}
+		} catch (error) {
+			next(error);
 		}
-		const result = await Expense.create({
-			name: req.body.name,
-			quantity: req.body.quantity,
-			description: req.body.description,
-		});
-		res.status(200).send(result);
 	};
 
 	const get = async (req: Request, res: Response, next: NextFunction) => {
@@ -30,28 +57,65 @@ const ExpenseController = () => {
 		res.status(200).send(result);
 	};
 
-	const update = async (req: Request, res: Response, next: NextFunction) => {
-		const result = await Expense.findById(req.params.id);
-		if (!result) {
-			res.status(400);
-			throw new Error("Expense not found");
+	const update = async (
+		req: AuthRequest,
+		res: Response,
+		next: NextFunction
+	) => {
+		try {
+			if (req.user) {
+				const result = await Expense.findById(req.params.id);
+				if (!result) {
+					res.status(400);
+					throw new Error("Expense not found");
+				}
+				// TODO: Find a way to check for expense ownership in middleware or extract function
+				if (result.user.toString() !== req.user.id) {
+					res.status(400);
+					throw new Error("User invalid");
+				}
+
+				const updated = await Expense.findByIdAndUpdate(
+					req.params.id,
+					req.body,
+					{ new: true }
+				);
+				res.status(200).send(updated);
+			} else {
+				res.status(400);
+				throw new Error("User invalid");
+			}
+		} catch (error) {
+			next(error);
 		}
-		const updated = await Expense.findByIdAndUpdate(
-			req.params.id,
-			req.body,
-			{ new: true }
-		);
-		res.status(200).send(updated);
 	};
 
-	const destroy = async (req: Request, res: Response, next: NextFunction) => {
-		const result = await Expense.findById(req.params.id);
-		if (!result) {
-			res.status(400);
-			throw new Error("Expense not found");
+	const destroy = async (
+		req: AuthRequest,
+		res: Response,
+		next: NextFunction
+	) => {
+		try {
+			if (req.user) {
+				const result = await Expense.findById(req.params.id);
+				if (!result) {
+					res.status(400);
+					throw new Error("Expense not found");
+				}	
+				// TODO: Find a way to check for expense ownership in middleware or extract function
+				if (result.user.toString() !== req.user.id) {
+					res.status(400);
+					throw new Error("User invalid");
+				}
+				const deleted = await Expense.findByIdAndDelete(req.params.id);
+				res.status(200).send(deleted);
+			} else {
+				res.status(400);
+				throw new Error("User invalid");
+			}
+		} catch (error) {
+			next(error);
 		}
-		const deleted = await Expense.findByIdAndDelete(req.params.id)
-		res.status(200).send(deleted);
 	};
 
 	return {
